refactor: extract 5 sub-configs from ConfigManager (1135→811 lines)
Create domain-specific config classes with forwarding properties
in ConfigManager for backward compatibility:

- AiConfig (179 lines): GPT/Gemini API keys, models, prompts, actions
- TtsConfig (95 lines): TTS locale, speed, type, voice items
- TranslationConfig (75 lines): translation language, mode, text style
- TouchConfig (99 lines): touch areas, gestures, multitouch, pagination
- DisplayConfig (114 lines): fonts, dark mode, zoom, highlights, PDF

ConfigManager exposes sub-configs as val ai/tts/translation/touch/display
and keeps forwarding properties so all 47 consumers compile unchanged.
